Add a `Makefile` to make running the app easier
Prior to this change, running the application had to be done manually.
Since this involves cd-ing into a new directory, its much nicer to have
this as an automated command.

This change adds a basic `Makefile` that has a `run-develop-native`
command which will spin up the application with uvicorns `--reload` flag
enabled.