Add the basic layout of the application
Prior to this change, there weren't any actual python files for this
application!

This change adds the basic files and starts the application layout for a
basic starlette application. The layout has been created with
extendability in mind, for instance it has a `config.py` module and splits
out the starlette routes and middleware into their own modules.