Add `conftest.py` to update the syspath
Prior to this change, python files inside of the tests directory
couldn't import the `triangler` module.

This change updates the syspath by means of the `conftest.py` file at
the root of the tests directory. This enables all tests under this path
to be able to import the `triangler` module.