Update `config.py` module for mandatory vars
Prior to this change, there was no way to enforce that a value is
mandatory in the config files.

This change adds some methods that allow us to properly type certain
config values as not being `None`. This solves a bunch of typing errors
and also gives us some runtime ensurance that certain config values
have been set.