Update `config.py` to use mandatory getters
Prior to this change, some of the config values that needed to be
mandatory were being gotten from the environment in a way that allowed
them to be set as `None`.

This change updates the `SQLALCHEMY_DATABASE_URL` and `SECRET_KEY`
config values to be gotten in a way where they cannot result in being
`None`, using the new mandatory getters.