Add `persistence.py` for creating datbase objects
Prior to this change, there were no connections being made to the
database.

This change adds the persistence module which contains code which sets
up engines and sessions to a given database url.

This change also updates the `.gitignore` file to ignore database files
so they will not be pushed upstream.